Refactor traits type signature
Refactor the type signature of the functions within the traits related
to compressed bytes, to output the new offset that will be used in
subsequent parsing or serialization calls.

Additionally, refactor the `len` methods of `CharacterString` and
`DomainLabel` to change their return types to `u8`. This is to make it
easier for callers of the function, as the length of these types must
always fit in a single byte, since the length of a character string is
always serialized as the first byte in the serialized bytes