Refactor trait signatures and implement parse
Refactor the signatures of `BytesSerializable` and
`CompressedBytesSerializable` to remove the associated error type
`ParseError`. This is to make it easier for implementations to handle
the `?` error propagation.

It is also not so necessary for callers to know where exactly the
parsing failed, as all they need to know is that it failed and handle it
accordingly. This info is only useful for debugging, of which there are
other ways to find out.

This commit also adds an implementation of `CharacterString::parse`